+++ apps/example-game/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-game
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-text-game/core@0.9.1
+
 ## 0.0.4
 
 ### Patch Changes