+++ apps/ui-test-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ui-test-app
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-text-game/core@0.9.1
+
 ## 0.1.4
 
 ### Patch Changes